GraphRAG/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# LangGraph & LangChain
langgraph==0.0.26
langchain==0.1.0
langchain-community==0.0.13
langchain-openai==0.0.2

# Vector Databases
chromadb==0.4.18
faiss-cpu==1.7.4

# Document Processing
pypdf2==3.0.1
python-docx==1.1.0
beautifulsoup4==4.12.2
markdown==3.5.1

# Text Processing & Embeddings
sentence-transformers==2.2.2
tiktoken==0.5.2
rank-bm25==0.2.2

# Async & Caching
redis==5.0.1
aioredis==2.0.1
httpx==0.25.2

# Monitoring & Logging
structlog==23.2.0
prometheus-client==0.19.0

# Utilities
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
